Moved validators to Core
PCOR-xxx

- Validator classes might become an interface in the future so we can
switch to different validation rules which could be defined from a file
or server-side
