+++ core/entity/CMakeLists.txt
@@ -1,8 +1,5 @@
 project (entity)
 
-# Includes root directory
-include_directories(.)
-
 # entity lib
 add_library (
     entity
@@ -17,15 +14,6 @@ add_library (
     personalid.hpp
     user.hpp
     user.cpp
-    # Validators
-    validator/addressvalidator.hpp
-    validator/addressvalidator.cpp
-    validator/contactdetailsvalidator.hpp
-    validator/contactdetailsvalidator.cpp
-    validator/personalidvalidator.hpp
-    validator/personalidvalidator.cpp
-    validator/personvalidator.hpp
-    validator/personvalidator.cpp
 )
 
 target_compile_options(entity PUBLIC "-fPIC")
