+++ core/validator/CMakeLists.txt
@@ -1,0 +1,18 @@
+project (validator)
+
+# validator lib
+add_library (
+    validator
+    STATIC
+    # Validators
+    addressvalidator.hpp
+    addressvalidator.cpp
+    contactdetailsvalidator.hpp
+    contactdetailsvalidator.cpp
+    personalidvalidator.hpp
+    personalidvalidator.cpp
+    personvalidator.hpp
+    personvalidator.cpp
+)
+
+target_compile_options(validator PUBLIC "-fPIC")
