+++ core/validator/contactdetailsvalidator.cpp
